test_gRPC_service/service_chunking/requirements.txt
fastmcp
grpcio
grpcio-tools
protobuf
llama-index
langchain
langchain-community
langchain-openai
pydantic
python-dotenv
asyncio
uvloop
redis
pymongo
neo4j
pymilvus
